refactor(linter): extract shared no-unsanitized sanitization analysis
