+++ crates/oxc_linter/src/rules/no_unsanitized/property.rs
@@ -1,20 +1,20 @@
 use oxc_ast::{
     AstKind,
-    ast::{
-        AssignmentOperator, AssignmentTarget, Expression, IdentifierReference,
-        VariableDeclarationKind,
-    },
+    ast::{AssignmentOperator, AssignmentTarget, MemberExpression},
 };
 use oxc_diagnostics::OxcDiagnostic;
 use oxc_macros::declare_oxc_lint;
-use oxc_semantic::SymbolId;
-use oxc_span::{GetSpan, Span};
+use oxc_span::Span;
 use oxc_str::CompactStr;
-use rustc_hash::FxHashSet;
 use schemars::JsonSchema;
 use serde::Deserialize;
 
-use crate::{AstNode, context::LintContext, rule::Rule};
+use crate::{
+    AstNode,
+    context::LintContext,
+    rule::Rule,
+    utils::{EscapeConfig, EscapeSchema, Sanitization, SinkValue},
+};
 
 fn unsafe_assignment_diagnostic(property: &str, span: Span) -> OxcDiagnostic {
     OxcDiagnostic::warn(format!("Unsafe assignment to {property}"))
@@ -24,33 +24,11 @@ fn unsafe_assignment_diagnostic(property: &str, span: Span) -> OxcDiagnostic {
         .with_label(span)
 }
 
-/// Escaping functions which mark a value as safe.
-#[derive(Debug, Clone)]
-struct Escape {
-    tagged_templates: Vec<CompactStr>,
-    methods: Vec<CompactStr>,
-}
-
-impl Default for Escape {
-    fn default() -> Self {
-        Self {
-            tagged_templates: vec![
-                CompactStr::new("Sanitizer.escapeHTML"),
-                CompactStr::new("escapeHTML"),
-            ],
-            methods: vec![
-                CompactStr::new("Sanitizer.unwrapSafeHTML"),
-                CompactStr::new("unwrapSafeHTML"),
-            ],
-        }
-    }
-}
-
 #[derive(Debug, Clone)]
 pub struct PropertyConfig {
     /// Property names which are treated as HTML sinks.
     properties: Vec<CompactStr>,
-    escape: Escape,
+    escape: EscapeConfig,
     variable_tracing: bool,
 }
 
@@ -58,7 +36,7 @@ impl Default for PropertyConfig {
     fn default() -> Self {
         Self {
             properties: vec![CompactStr::new("innerHTML"), CompactStr::new("outerHTML")],
-            escape: Escape::default(),
+            escape: EscapeConfig::default(),
             variable_tracing: true,
         }
     }
@@ -66,17 +44,6 @@ impl Default for PropertyConfig {
 
 #[derive(Debug, Default, Clone)]
 pub struct Property(Box<PropertyConfig>);
-
-#[derive(Debug, Default, Deserialize, JsonSchema)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-struct EscapeSchema {
-    /// Tagged template functions which return safe HTML.
-    /// Defaults to `["Sanitizer.escapeHTML", "escapeHTML"]`.
-    tagged_templates: Option<Vec<String>>,
-    /// Methods which return safe HTML.
-    /// Defaults to `["Sanitizer.unwrapSafeHTML", "unwrapSafeHTML"]`.
-    methods: Option<Vec<String>>,
-}
 
 #[derive(Debug, Default, Deserialize, JsonSchema)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
@@ -159,15 +126,8 @@ impl Rule for Property {
         }
         let options: PropertyOptionsSchema = serde_json::from_value(options)?;
 
-        if let Some(escape) = options.escape {
-            if let Some(tagged_templates) = escape.tagged_templates {
-                config.escape.tagged_templates =
-                    tagged_templates.iter().map(|s| CompactStr::from(s.as_str())).collect();
-            }
-            if let Some(methods) = escape.methods {
-                config.escape.methods =
-                    methods.iter().map(|s| CompactStr::from(s.as_str())).collect();
-            }
+        if let Some(escape) = &options.escape {
+            config.escape.apply(escape);
         }
         if let Some(variable_tracing) = options.variable_tracing {
             config.variable_tracing = variable_tracing;
@@ -193,127 +153,11 @@ impl Rule for Property {
             return;
         }
 
-        let mut seen = FxHashSet::default();
-        if !self.is_allowed_expression(&assignment.right, ctx, &mut seen) {
+        let sanitization =
+            Sanitization { escape: &self.0.escape, variable_tracing: self.0.variable_tracing };
+        if !sanitization.is_allowed(SinkValue::Expression(&assignment.right), ctx) {
             ctx.diagnostic(unsafe_assignment_diagnostic(property, assignment.span));
         }
-    }
-}
-
-impl Property {
-    fn is_allowed_expression<'a>(
-        &self,
-        expression: &Expression<'a>,
-        ctx: &LintContext<'a>,
-        seen: &mut FxHashSet<SymbolId>,
-    ) -> bool {
-        match expression {
-            // A literal cannot carry attacker controlled markup, only malice.
-            Expression::StringLiteral(_)
-            | Expression::NumericLiteral(_)
-            | Expression::BooleanLiteral(_)
-            | Expression::NullLiteral(_)
-            | Expression::BigIntLiteral(_)
-            | Expression::RegExpLiteral(_) => true,
-            // Only the `${...}` parts need checking, a quasi is raw text.
-            // A template literal without interpolations is therefore always safe.
-            Expression::TemplateLiteral(template) => template
-                .expressions
-                .iter()
-                .all(|expression| self.is_allowed_expression(expression, ctx, seen)),
-            Expression::TaggedTemplateExpression(tagged) => {
-                is_allowed_callee(&tagged.tag, &self.0.escape.tagged_templates, ctx)
-            }
-            Expression::CallExpression(call) => {
-                is_allowed_callee(&call.callee, &self.0.escape.methods, ctx)
-            }
-            Expression::BinaryExpression(binary) => {
-                self.is_allowed_expression(&binary.left, ctx, seen)
-                    && self.is_allowed_expression(&binary.right, ctx, seen)
-            }
-            Expression::ParenthesizedExpression(paren) => {
-                self.is_allowed_expression(&paren.expression, ctx, seen)
-            }
-            Expression::TSAsExpression(_)
-            | Expression::TSSatisfiesExpression(_)
-            | Expression::TSNonNullExpression(_)
-            | Expression::TSTypeAssertion(_)
-            | Expression::TSInstantiationExpression(_) => {
-                self.is_allowed_expression(expression.get_inner_expression(), ctx, seen)
-            }
-            Expression::Identifier(identifier) => self.is_allowed_identifier(identifier, ctx, seen),
-            _ => false,
-        }
-    }
-
-    /// Traces an identifier back to its declaration.
-    ///
-    /// Only `let`/`const` bindings whose initializer and every write reference are
-    /// themselves allowed expressions are considered safe.
-    fn is_allowed_identifier<'a>(
-        &self,
-        identifier: &IdentifierReference<'a>,
-        ctx: &LintContext<'a>,
-        seen: &mut FxHashSet<SymbolId>,
-    ) -> bool {
-        if !self.0.variable_tracing {
-            return false;
-        }
-        let scoping = ctx.scoping();
-        // Unresolved references (globals, implicit assignments) cannot be traced.
-        let Some(symbol_id) = scoping.get_reference(identifier.reference_id()).symbol_id() else {
-            return false;
-        };
-        // Guard against cycles such as `let a = ''; a = `${a}<p>`;`.
-        if !seen.insert(symbol_id) {
-            return false;
-        }
-
-        let allowed = self.is_allowed_symbol(symbol_id, ctx, seen);
-        seen.remove(&symbol_id);
-        allowed
-    }
-
-    fn is_allowed_symbol(
-        &self,
-        symbol_id: SymbolId,
-        ctx: &LintContext<'_>,
-        seen: &mut FxHashSet<SymbolId>,
-    ) -> bool {
-        let scoping = ctx.scoping();
-        let declaration = ctx.nodes().get_node(scoping.symbol_declaration(symbol_id));
-        let AstKind::VariableDeclarator(declarator) = declaration.kind() else {
-            // Function parameters, imports, classes, ... are not traceable.
-            return false;
-        };
-        // `var` can be overwritten in ways that are not visible here.
-        if !matches!(declarator.kind, VariableDeclarationKind::Let | VariableDeclarationKind::Const)
-        {
-            return false;
-        }
-        if let Some(init) = &declarator.init
-            && !self.is_allowed_expression(init, ctx, seen)
-        {
-            return false;
-        }
-
-        scoping.get_resolved_references(symbol_id).filter(|reference| reference.is_write()).all(
-            |reference| {
-                write_expression(ctx, reference.node_id())
-                    .is_some_and(|expression| self.is_allowed_expression(expression, ctx, seen))
-            },
-        )
-    }
-}
-
-/// The right-hand side of the assignment a write reference belongs to.
-fn write_expression<'a, 'b>(
-    ctx: &'b LintContext<'a>,
-    node_id: oxc_semantic::NodeId,
-) -> Option<&'b Expression<'a>> {
-    match ctx.nodes().parent_kind(node_id) {
-        AstKind::AssignmentExpression(assignment) => Some(&assignment.right),
-        _ => None,
     }
 }
 
@@ -342,36 +186,7 @@ fn assigned_property_name<'a>(target: &AssignmentTarget<'a>) -> Option<&'a str> 
         None => simple.as_member_expression(),
     }?;
     match member {
-        oxc_ast::ast::MemberExpression::StaticMemberExpression(member) => {
-            Some(member.property.name.as_str())
-        }
-        _ => None,
-    }
-}
-
-/// `foo` for `foo()`, `obj.foo` for `obj.foo()`, using source text for
-/// non-identifier objects, matching the upstream `getCodeName` helper.
-fn is_allowed_callee<'a>(
-    callee: &Expression<'a>,
-    allowed: &[CompactStr],
-    ctx: &LintContext<'a>,
-) -> bool {
-    let Some(name) = callee_code_name(callee, ctx) else {
-        return false;
-    };
-    allowed.iter().any(|candidate| candidate.as_str() == name)
-}
-
-fn callee_code_name<'a>(callee: &Expression<'a>, ctx: &LintContext<'a>) -> Option<String> {
-    match callee {
-        Expression::Identifier(identifier) => Some(identifier.name.to_string()),
-        Expression::StaticMemberExpression(member) => {
-            let object = match &member.object {
-                Expression::Identifier(identifier) => identifier.name.to_string(),
-                object => ctx.source_range(object.span()).to_string(),
-            };
-            Some(format!("{object}.{}", member.property.name))
-        }
+        MemberExpression::StaticMemberExpression(member) => Some(member.property.name.as_str()),
         _ => None,
     }
 }
